fix: relax validation of disallowedCategories prop of dapp-explorer FF payload
